+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Change Log
 ==========
 
+2016-08-26
+----------
+
+Changed:
+
+ * allow fetching rules by class name, not just instance
+
 2016-08-06
 ----------